refactor(cache): extract remote cache client
