+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 resolver = "3"
 members = [
+  "crates/mise-cache-core",
   "crates/vfox",
   "crates/aqua-registry",
   "crates/mise-interactive-config",
@@ -114,6 +115,7 @@ filetime = "0.2"
 flate2 = "1"
 futures-util = "0.3.32"
 mise-sigstore = { path = "crates/mise-sigstore", default-features = false }
+mise-cache-core = { path = "crates/mise-cache-core", default-features = false }
 fslock = "0.2.1"
 gix = { version = "<1", features = ["worktree-mutation"] }
 glob = "0.3"
@@ -286,6 +288,7 @@ native-tls = [
   "rattler/native-tls",
   "rattler_repodata_gateway/native-tls",
   "mise-sigstore/native-tls",
+  "mise-cache-core/native-tls",
   "ubi/native-tls",
   "vfox/native-tls",
   "xx/native-tls",
@@ -297,6 +300,7 @@ rustls = [
   "rattler/rustls",
   "rattler_repodata_gateway/rustls",
   "mise-sigstore/rustls",
+  "mise-cache-core/rustls",
   "ubi/rustls-tls",
   "vfox/rustls",
   "xx/rustls",
@@ -308,6 +312,7 @@ rustls-native-roots = [
   "rattler/rustls",
   "rattler_repodata_gateway/rustls",
   "mise-sigstore/rustls",
+  "mise-cache-core/rustls-native-roots",
   "ubi/rustls-tls",
   "vfox/rustls-native-roots",
   "xx/rustls-native-roots",
